Moved sharable classes to Common library
This commit also contains some improvements to general business logic
