+++ src/MediaFileManager/MediaFileManager.Common/Models/Audio/Album.cs
@@ -1,0 +1,19 @@
+﻿namespace MediaFileManager.Common.Models.Audio;
+
+public class Album : BindableBase
+{
+    private string name;
+    private List<Song> songs;
+
+    public string Name
+    {
+        get => name;
+        set => SetProperty(ref name, value);
+    }
+
+    public List<Song> Songs
+    {
+        get => songs;
+        set => SetProperty(ref songs, value);
+    }
+}
