+++ src/MediaFileManager/MediaFileManager.Common/Models/OutputMessage.cs
@@ -1,0 +1,9 @@
+﻿using System.Drawing;
+
+namespace MediaFileManager.Common.Models;
+
+public class OutputMessage
+{
+    public string Message { get; set; }
+    public Color MessageColor { get; set; }
+}
