+++ src/MediaFileManager/MediaFileManager.Common/Models/OutputMessageLevel.cs
@@ -1,0 +1,11 @@
+﻿namespace MediaFileManager.Common.Models;
+
+public enum OutputMessageLevel
+{
+    Normal,
+    Informational,
+    Success,
+    Warning,
+    Error
+}
+
