+++ src/MediaFileManager/MediaFileManager.Common/Models/TagWorkerParameters.cs
@@ -1,7 +1,6 @@
-﻿using MediaFileManager.Desktop.Models.AudioBook;
-using System.Collections.Generic;
+﻿using MediaFileManager.Common.Models.AudioBook;
 
-namespace MediaFileManager.Desktop.Models
+namespace MediaFileManager.Common.Models
 {
     public class TagWorkerParameters
     {
