+++ src/MediaFileManager/MediaFileManager.Common/Models/WorkerResult.cs
@@ -1,0 +1,7 @@
+﻿namespace MediaFileManager.Common.Models;
+
+public class WorkerResult
+{
+    public bool IsPreview { get; set; }
+    public string FinalMessage { get; set; }
+}
