+++ src/MediaFileManager/MediaFileManager.Desktop/Helpers/PersonalizationHelpers.cs
@@ -2,63 +2,66 @@
 using System.Collections.Generic;
 using System.Windows;
 
-namespace MediaFileManager.Desktop.Helpers
+namespace MediaFileManager.Desktop.Helpers;
+
+/// <summary>
+/// A helper class to help with changing themes via ResourceDictionary swap.
+/// </summary>
+public static class PersonalizationHelpers
 {
-    public static class PersonalizationHelpers
+    /// <summary>
+    /// Helper for merging Telerik theme into the App Resources
+    /// </summary>
+    /// <param name="themeAssemblyName">Name of the theme to use (e.g. "Fluent" or "Office2013"). Query this class's ThemeNames list for available names to use.</param>
+    public static void UpdateTheme(string themeAssemblyName)
     {
-        /// <summary>
-        /// Helper for merging Telerik theme into the App Resources
-        /// </summary>
-        /// <param name="themeAssemblyName">Name of the theme to use (e.g. "Fluent" or "Office2013"). Query this class's ThemeNames list for available names to use.</param>
-        public static void UpdateTheme(string themeAssemblyName)
+        Application.Current.Resources.MergedDictionaries.Clear();
+
+        foreach (var fileName in ResourceDictionaryList)
         {
-            Application.Current.Resources.MergedDictionaries.Clear();
-
-            foreach (var fileName in ResourceDictionaryList)
-            {
-                Application.Current.Resources.MergedDictionaries.Add(new ResourceDictionary
-                {
-                    Source = new Uri($"/Telerik.Windows.Themes.{themeAssemblyName};component/Themes/{fileName}", UriKind.RelativeOrAbsolute)
-                });
-            }
-
             Application.Current.Resources.MergedDictionaries.Add(new ResourceDictionary
             {
-                Source = new Uri("/Styles/GeneralStyles.xaml", UriKind.RelativeOrAbsolute)
+                Source = new Uri($"/Telerik.Windows.Themes.{themeAssemblyName};component/Themes/{fileName}", UriKind.RelativeOrAbsolute)
             });
         }
 
-        /// <summary>
-        /// Helper for merging Telerik themes into a pre-defined ResourceDictionary. This is useful for custom controls that do not have access to the App resources
-        /// </summary>
-        /// <param name="source">The ResourceDictionary to merge the Telerik theme ResourceDictionaries into.</param>
-        /// <param name="themeAssemblyName">Name of the theme to use (e.g. "Fluent" or "Office2013"). Query this class's ThemeNames list for available names to use.</param>
-        public static void UpdateTheme(ResourceDictionary source, string themeAssemblyName)
+        Application.Current.Resources.MergedDictionaries.Add(new ResourceDictionary
         {
-            if(source == null)
-            {
-                throw new Exception("You must pass a valid ResourceDictionary to reset the themes.");
-            }
+            Source = new Uri("/Styles/GeneralStyles.xaml", UriKind.RelativeOrAbsolute)
+        });
+    }
 
-            source.MergedDictionaries.Clear();
+    /// <summary>
+    /// Helper for merging Telerik themes into a pre-defined ResourceDictionary. This is useful for custom controls that do not have access to the App resources
+    /// </summary>
+    /// <param name="source">The ResourceDictionary to merge the Telerik theme ResourceDictionaries into.</param>
+    /// <param name="themeAssemblyName">Name of the theme to use (e.g. "Fluent" or "Office2013"). Query this class's ThemeNames list for available names to use.</param>
+    public static void UpdateTheme(ResourceDictionary source, string themeAssemblyName)
+    {
+        if (source == null)
+        {
+            throw new Exception("You must pass a valid ResourceDictionary to reset the themes.");
+        }
 
-            foreach (var fileName in ResourceDictionaryList)
-            {
-                source.MergedDictionaries.Add(new ResourceDictionary
-                {
-                    Source = new Uri($"/Telerik.Windows.Themes.{themeAssemblyName};component/Themes/{fileName}", UriKind.RelativeOrAbsolute)
-                });
-            }
+        source.MergedDictionaries.Clear();
 
+        foreach (var fileName in ResourceDictionaryList)
+        {
             source.MergedDictionaries.Add(new ResourceDictionary
             {
-                Source = new Uri("/Styles/GeneralStyles.xaml", UriKind.RelativeOrAbsolute)
+                Source = new Uri($"/Telerik.Windows.Themes.{themeAssemblyName};component/Themes/{fileName}", UriKind.RelativeOrAbsolute)
             });
         }
 
-        // This is the ensure we are merging the correct Resources
-        private static readonly string[] ResourceDictionaryList =
+        source.MergedDictionaries.Add(new ResourceDictionary
         {
+            Source = new Uri("/Styles/GeneralStyles.xaml", UriKind.RelativeOrAbsolute)
+        });
+    }
+
+    // This is the ensure we are merging the correct Resources
+    private static readonly string[] ResourceDictionaryList =
+    {
             "System.Windows.xaml",
             "Telerik.Windows.Controls.xaml",
             "Telerik.Windows.Controls.Data.xaml",
@@ -70,25 +73,24 @@ namespace MediaFileManager.Desktop.Helpers
             "Telerik.Windows.Controls.RichTextBox.xaml"
         };
 
-        /// <summary>
-        /// List of themes names that are currently supported.
-        /// </summary>
-        public static List<string> ThemeNames { get; } = new List<string>
+    /// <summary>
+    /// List of themes names that are currently supported.
+    /// </summary>
+    public static List<string> ThemeNames { get; } = new List<string>
         {
-            "Fluent", 
-            "VisualStudio2019", 
-            "Crystal", 
-            "Expression_Dark", 
+            "Fluent",
+            "VisualStudio2019",
+            "Crystal",
+            "Expression_Dark",
             "Green",
-            "Material", 
-            "Office2013", 
-            "Office2016", 
-            "Summer", 
+            "Material",
+            "Office2013",
+            "Office2016",
+            "Summer",
             "Transparent",
-            "Vista", 
-            "VisualStudio2013", 
-            "Windows7", 
+            "Vista",
+            "VisualStudio2013",
+            "Windows7",
             "Windows8",
         };
-    }
 }
