--- src/MediaFileManager/MediaFileManager.Desktop/Models/Audio/Artist.cs
@@ -1,22 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace MediaFileManager.Desktop.Models.Audio
-{
-    public class Artist : BindableBase
-    {
-        private string name;
-        private List<Album> albums;
-
-        public string Name
-        {
-            get => name;
-            set => SetProperty(ref name, value);
-        }
-
-        public List<Album> Albums 
-        { 
-            get => albums;
-            set => SetProperty(ref albums, value);
-        }
-    }
-}
