--- src/MediaFileManager/MediaFileManager.Desktop/Models/OutputMessage.cs
@@ -1,10 +1,0 @@
-﻿using System.Windows.Media;
-
-namespace MediaFileManager.Desktop.Models
-{
-    public class OutputMessage
-    {
-        public string Message { get; set; }
-        public Color MessageColor { get; set; }
-    }
-}
