--- src/MediaFileManager/MediaFileManager.Desktop/Models/OutputMessageLevel.cs
@@ -1,11 +1,0 @@
-﻿namespace MediaFileManager.Desktop.Models
-{
-    public enum OutputMessageLevel
-    {
-        Normal,
-        Informational,
-        Success,
-        Warning,
-        Error
-    }
-}
