--- src/MediaFileManager/MediaFileManager.Desktop/Models/WorkerProgress.cs
@@ -1,10 +1,0 @@
-﻿namespace MediaFileManager.Desktop.Models
-{
-    public class WorkerProgress
-    {
-        public int PercentComplete { get; set; }
-        public string BusyMessage { get; set; }
-        public string FileName { get; set; }
-        public bool IsPreview { get; set; }
-    }
-}
