--- src/MediaFileManager/MediaFileManager.Desktop/Models/WorkerResult.cs
@@ -1,8 +1,0 @@
-﻿namespace MediaFileManager.Desktop.Models
-{
-    public class WorkerResult
-    {
-        public bool IsPreview { get; set; }
-        public string FinalMessage { get; set; }
-    }
-}
