+++ src/MediaFileManager/MediaFileManager.Desktop/UserControls/AnimatedLogo.xaml.cs
@@ -2,19 +2,18 @@
 using System.Windows.Controls;
 using System.Windows.Media.Animation;
 
-namespace MediaFileManager.Desktop.UserControls
-{
-    public partial class AnimatedLogo : UserControl
-    {
-        public AnimatedLogo()
-        {
-            InitializeComponent();
-            Loaded += AnimatedLogo_Loaded;
-        }
+namespace MediaFileManager.Desktop.UserControls;
 
-        private void AnimatedLogo_Loaded(object sender, RoutedEventArgs e)
-        {
-            (Resources["GearsRotatingStoryboard"] as Storyboard)?.Begin();
-        }
+public partial class AnimatedLogo : UserControl
+{
+    public AnimatedLogo()
+    {
+        InitializeComponent();
+        Loaded += AnimatedLogo_Loaded;
+    }
+
+    private void AnimatedLogo_Loaded(object sender, RoutedEventArgs e)
+    {
+        (Resources["GearsRotatingStoryboard"] as Storyboard)?.Begin();
     }
 }
