+++ src/MediaFileManager/MediaFileManager.Desktop/Windows/MainWindow.xaml.cs
@@ -4,84 +4,82 @@ using System.Windows.Controls;
 using MediaFileManager.Desktop.Helpers;
 using MediaFileManager.Desktop.Views;
 
-namespace MediaFileManager.Desktop.Windows
+namespace MediaFileManager.Desktop.Windows;
+
+[System.Diagnostics.CodeAnalysis.SuppressMessage("Interoperability", "CA1416:Validate platform compatibility", Justification = "Irrelevant")]
+public partial class MainWindow : Window
 {
-    public partial class MainWindow : Window
+    public MainWindow()
     {
-        public MainWindow()
+        if (Properties.Settings.Default.PreferredTheme != "Fluent")
         {
-            if (Properties.Settings.Default.PreferredTheme != "Fluent")
-            {
-                PersonalizationHelpers.UpdateTheme(Properties.Settings.Default.PreferredTheme);
-            }
-
-            InitializeComponent();
-
-            ThemeComboBox.ItemsSource = PersonalizationHelpers.ThemeNames;
-            ThemeComboBox.SelectedItem = Properties.Settings.Default.PreferredTheme;
-
-            FileTypeComboBox.ItemsSource = new List<string> {"Videos", "Audiobooks", "Music"}; 
-            FileTypeComboBox.SelectedIndex = Properties.Settings.Default.SelectedViewIndex;
+            PersonalizationHelpers.UpdateTheme(Properties.Settings.Default.PreferredTheme);
         }
 
-        private void FileTypeComboBox_OnSelectionChanged(object sender, SelectionChangedEventArgs e)
+        InitializeComponent();
+
+        ThemeComboBox.ItemsSource = PersonalizationHelpers.ThemeNames;
+        ThemeComboBox.SelectedItem = Properties.Settings.Default.PreferredTheme;
+
+        FileTypeComboBox.ItemsSource = new List<string> { "Videos", "Audiobooks", "Music" };
+        FileTypeComboBox.SelectedIndex = Properties.Settings.Default.SelectedViewIndex;
+    }
+    
+    private void FileTypeComboBox_OnSelectionChanged(object sender, SelectionChangedEventArgs e)
+    {
+        if (e.AddedItems != null)
         {
-            if (e.AddedItems != null)
+            var selectedItem = FileTypeComboBox.SelectedItem.ToString();
+
+            if (selectedItem == "Videos")
             {
-                var selectedItem = FileTypeComboBox.SelectedItem.ToString();
+                SelectedViewBorder.Child = new VideoFilesView();
+            }
+            else if (selectedItem == "Audiobooks")
+            {
+                SelectedViewBorder.Child = new AudiobookFilesView();
+            }
+            else if (selectedItem == "Music")
+            {
+                SelectedViewBorder.Child = new MusicFilesView();
+            }
 
-                if (selectedItem == "Videos")
-                {
-                    SelectedViewBorder.Child = new VideoFilesView();
-                }
-                else if (selectedItem == "Audiobooks")
-                {
-                    SelectedViewBorder.Child = new AudiobookFilesView();
-                }
-                else if (selectedItem == "Music")
-                {
-                    SelectedViewBorder.Child = new MusicFilesView();
-                }
+            Title = $"Media File Manager - {selectedItem}";
 
-                Title = $"Media File Manager - {selectedItem}";
-
-                Properties.Settings.Default.SelectedViewIndex = FileTypeComboBox.SelectedIndex;
-
-                Microsoft.AppCenter.Analytics.Analytics.TrackEvent("View Selected", new Dictionary<string, string>
+            Properties.Settings.Default.SelectedViewIndex = FileTypeComboBox.SelectedIndex;
+            
+            Microsoft.AppCenter.Analytics.Analytics.TrackEvent("View Selected", new Dictionary<string, string>
                 {
                     { "FileType", selectedItem }
                 });
-            }
         }
-
-        private void ThemeComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
+    }
+    
+    private void ThemeComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
+    {
+        if (e.AddedItems != null)
         {
-            if (e.AddedItems != null)
-            {
-                var selectedThemeName = ThemeComboBox.SelectedItem.ToString();
+            var selectedThemeName = ThemeComboBox.SelectedItem.ToString();
 
-                PersonalizationHelpers.UpdateTheme(selectedThemeName);
-
-                Microsoft.AppCenter.Analytics.Analytics.TrackEvent("ThemeChanged", new Dictionary<string, string>
+            PersonalizationHelpers.UpdateTheme(selectedThemeName);
+            
+            Microsoft.AppCenter.Analytics.Analytics.TrackEvent("ThemeChanged", new Dictionary<string, string>
                 {
                     { "Theme Name", selectedThemeName }
                 });
 
-                Properties.Settings.Default.PreferredTheme = selectedThemeName;
-                Properties.Settings.Default.Save();
-            } 
+            Properties.Settings.Default.PreferredTheme = selectedThemeName;
+            Properties.Settings.Default.Save();
         }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Interoperability", "CA1416:Validate platform compatibility", Justification = "Irrelevant")]
-        private void HelpButton_OnClick(object sender, RoutedEventArgs e)
-        {
-            new HelpWindow().Show();
-        }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Interoperability", "CA1416:Validate platform compatibility", Justification = "Irrelevant")]
-        private void AboutButton_OnClick(object sender, RoutedEventArgs e)
-        {
-            new AboutWindow().Show();
-        }
+    }
+    
+    private void HelpButton_OnClick(object sender, RoutedEventArgs e)
+    {
+        new HelpWindow().Show();
+    }
+    
+    private void AboutButton_OnClick(object sender, RoutedEventArgs e)
+    {
+        new AboutWindow().Show();
     }
 }
