+++ src/MediaFileManager/MediaFileManager.sln
@@ -7,16 +7,22 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MediaFileManager.Desktop", 
 EndProject
 Project("{C7167F0D-BC9F-4E6E-AFE1-012C56B48DB5}") = "PackageProject", "PackageProject\PackageProject.wapproj", "{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MediaFileManager.Common", "MediaFileManager.Common\MediaFileManager.Common.csproj", "{4F57613F-CFB0-4237-ADEB-80B09F67643F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
 		Debug|ARM64 = Debug|ARM64
 		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
 		Release|ARM64 = Release|ARM64
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Debug|Any CPU.Build.0 = Debug|x64
 		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Debug|ARM64.ActiveCfg = Debug|arm64
 		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Debug|ARM64.Build.0 = Debug|arm64
 		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Debug|ARM64.Deploy.0 = Debug|arm64
@@ -26,6 +32,8 @@ Global
 		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Debug|x86.ActiveCfg = Debug|x86
 		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Debug|x86.Build.0 = Debug|x86
 		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Debug|x86.Deploy.0 = Debug|x86
+		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Release|Any CPU.ActiveCfg = Release|x64
+		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Release|Any CPU.Build.0 = Release|x64
 		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Release|ARM64.ActiveCfg = Release|arm64
 		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Release|ARM64.Build.0 = Release|arm64
 		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Release|ARM64.Deploy.0 = Release|arm64
@@ -35,6 +43,9 @@ Global
 		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Release|x86.ActiveCfg = Release|x86
 		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Release|x86.Build.0 = Release|x86
 		{48BFA162-9C25-4D80-AD49-A0B5795F4798}.Release|x86.Deploy.0 = Release|x86
+		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Debug|Any CPU.Build.0 = Debug|x64
+		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Debug|Any CPU.Deploy.0 = Debug|x64
 		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Debug|ARM64.ActiveCfg = Debug|ARM64
 		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Debug|ARM64.Build.0 = Debug|ARM64
 		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Debug|ARM64.Deploy.0 = Debug|ARM64
@@ -44,6 +55,9 @@ Global
 		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Debug|x86.ActiveCfg = Debug|x64
 		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Debug|x86.Build.0 = Debug|x64
 		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Debug|x86.Deploy.0 = Debug|x64
+		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Release|Any CPU.ActiveCfg = Release|x64
+		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Release|Any CPU.Build.0 = Release|x64
+		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Release|Any CPU.Deploy.0 = Release|x64
 		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Release|ARM64.ActiveCfg = Release|ARM64
 		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Release|ARM64.Build.0 = Release|ARM64
 		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Release|ARM64.Deploy.0 = Release|ARM64
@@ -53,6 +67,22 @@ Global
 		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Release|x86.ActiveCfg = Release|x86
 		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Release|x86.Build.0 = Release|x86
 		{79EC2C31-3930-4C0B-AB9F-E4EA6BC8800B}.Release|x86.Deploy.0 = Release|x86
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Debug|x64.Build.0 = Debug|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Debug|x86.Build.0 = Debug|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Release|ARM64.Build.0 = Release|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Release|x64.ActiveCfg = Release|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Release|x64.Build.0 = Release|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Release|x86.ActiveCfg = Release|Any CPU
+		{4F57613F-CFB0-4237-ADEB-80B09F67643F}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
